Move status type
